requirements.txt
numpy
pandas
notebook
mutagen # for extracting information from files
jupyter_contrib_nbextensions
tqdm
absl-py
tinytag
pydub